Update AdminVM and AdminJobs
Add ICommands and Relaycommands to Add, Delete and Update Jobs.

Delete button is disabled if a job is not selected

New button created a placeholder SelectedJob to get the data from the form and to use for adding it to he database

Update commands gets the details of the selected job and updates the database

+++ BayWyn_Couriers/Models/Job.cs
@@ -21,7 +21,7 @@ namespace BayWyn_Couriers.Models
         public DateTime EndDate { get; set; }
 
         public string Description { get; set; }
-        public float Cost { get; set; }
+        public double Cost { get; set; }
         public string JobStatus { get; set; }
     }
 }
